Use tempalted form creation
Currently, form factories store creation parameters and know what is
needed on a per-form basis for that form's creation.

This isn't strictly necessary for our purposes, though. Rather than a
templated form factory, we could templatize the form creation function.

This commit does that, removing the superfluous form factories.
